backend/requirements.txt
fastapi
pydantic
openai
python-dotenv
uvicorn
python-multipart